% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power_laws.R
\name{plot_pl_cdf}
\alias{plot_pl_cdf}
\title{Plot a power law's cumulative density function (CDF)}
\usage{
plot_pl_cdf(x, ticks_x = 5, ticks_y = 5, xlab = "x", ylab = "P(X >= x)")
}
\arguments{
\item{x}{numeric; A vector of discrete or continuous values.}

\item{ticks_x}{integer; Number of ticks in the x axis.}

\item{ticks_y}{integer; Number of ticks in the y axis.}

\item{xlab}{character; Label for the x axis.}

\item{ylab}{character; Label for the y axis.}
}
\value{
A ggplot object with the plot.
}
\description{
Given a vector of discrete or continuous values, it fits a power law
distribution to the data and plots the CDF in log-log scale.
}
\examples{
# Use log-binning to plot CDFs for the generated power law distributions
p_disc <- plot_pl_cdf(rnd_disc_pl(10000, 2, 2.3))
p_cont <- plot_pl_cdf(rnd_cont_pl(10000, 2, 2.3))

}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
