% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power_laws.R
\name{plot_pl_pdf}
\alias{plot_pl_pdf}
\title{Plot a power law's probability density function (PDF)}
\usage{
plot_pl_pdf(x, b = 2, ticks_x = 5, ticks_y = 5, xlab = "x",
  ylab = "p(x)")
}
\arguments{
\item{x}{numeric; A vector of discrete or continuous values.}

\item{b}{numeric; Bins have size b^(n-1).}

\item{ticks_x}{integer; Number of ticks in the x axis.}

\item{ticks_y}{integer; Number of ticks in the y axis.}

\item{xlab}{character; Label for the x axis.}

\item{ylab}{character; Label for the y axis.}
}
\value{
A ggplot object with the plot.
}
\description{
Given a vector of discrete or continuous values, it fits a power law
distribution to the data and plots the PDF in log-log scale.
}
\examples{
# Use log-binning to plot PDFs for the generated power law distributions
p_disc <- plot_pl_pdf(rnd_disc_pl(10000, 2, 2.3), 2)
p_cont <- plot_pl_pdf(rnd_cont_pl(10000, 2, 2.3), 2)

}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
