% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power_laws.R
\name{rnd_cont_pl}
\alias{rnd_cont_pl}
\title{Random numbers from continuous power law distribution}
\usage{
rnd_cont_pl(n, xmin = 2, gma = 2.5)
}
\arguments{
\item{n}{integer; The number of values to generate.}

\item{xmin}{numeric; Start of the power law behaviour.}

\item{gma}{numeric; Exponent of the power law.}
}
\value{
A vector of n power-law distributed values.
}
\description{
Generate n power-law distributed continuous values.
}
\examples{
# Generate 10000 power-law distributed values
pl_cont <- rnd_cont_pl(10000, 2, 2.3)

}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
